refactor(session-sweep): move to utils and handle discarded result in index.ts
session-sweep.ts only depends on core types (TmuxSessionManagerCorePort,
Logger, Result, AutobeatError), so it belongs in utils/ not services/.
Core layer must not depend on services layer per Clean Architecture rules.

- git mv src/services/session-sweep.ts → src/utils/session-sweep.ts
- Update imports in container.ts and index.ts
- Add sweepTmuxSessions re-export in utils/index.ts
- In index.ts: capture Result and log on listSessions() failure (logger available)
- In container.ts: add INTENTIONAL comment explaining why result is discarded
  (no logger at that call site; must not block DB close)
